+++ click_extra/platforms.py
@@ -85,8 +85,16 @@ def is_hurd() -> bool:
 
 @cache
 def is_linux() -> bool:
-    """Return `True` only if current platform is of the Linux family."""
-    return sys.platform.startswith("linux")
+    """Return `True` only if current platform is of the Linux family.
+
+    Excludes WSL1 and WSL2 from this check to `avoid false positives
+    <https://github.com/kdeldycke/meta-package-manager/issues/944>`_.
+    """
+    return  (
+        sys.platform.startswith("linux")
+        and not is_wsl1()
+        and not is_wsl2()
+    )
 
 
 @cache